requirements.txt
psutil
requests
python-dotenv
beautifulsoup4
PyPDF2
Pillow
pytesseract
pydub
matplotlib
playwright
croniter
youtube-transcript-api
datasets
groq
pyttsx3
langchain-text-splitters
langchain-huggingface
langchain-community
langchain-core
faiss-cpu
sentence-transformers
colorama
rich
tabulate
